Combine the tiny voice-model/src/payload/*.rs files
