--- voice-model/src/payload/heartbeat.rs
@@ -1,12 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use crate::util::json_safe_u64;
-
-/// Used to keep the websocket connection alive.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Deserialize, Serialize)]
-#[serde(transparent)]
-pub struct Heartbeat {
-    /// Random number generated by the client, to be mirrored by the server.
-    #[serde(with = "json_safe_u64")]
-    pub nonce: u64,
-}
