--- voice-model/src/payload/heartbeat_ack.rs
@@ -1,12 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use crate::util::json_safe_u64;
-
-/// Heartbeat ACK, received by the client to show the server's receipt of a heartbeat.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Deserialize, Serialize)]
-#[serde(transparent)]
-pub struct HeartbeatAck {
-    /// Random 64-bit number previously generated by the client, mirrored by the server.
-    #[serde(with = "json_safe_u64")]
-    pub nonce: u64,
-}
