--- voice-model/src/payload/identify.rs
@@ -1,18 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use crate::id::*;
-
-/// Used to begin a voice websocket connection.
-#[derive(Clone, Debug, Eq, Hash, PartialEq, Deserialize, Serialize)]
-pub struct Identify {
-    /// GuildId which the target voice channel belongs to.
-    pub server_id: GuildId,
-    /// Authentication session received from Discord's main gateway as part of a
-    /// `"VOICE_STATE_UPDATE"` message.
-    pub session_id: String,
-    /// Authentication token received from Discord's main gateway as part of a
-    /// `"VOICE_SERVER_UPDATE"` message.
-    pub token: String,
-    /// UserId of the client who is connecting.
-    pub user_id: UserId,
-}
