--- voice-model/src/payload/mod.rs
@@ -1,25 +1,0 @@
-//! Message bodies used in gateway event-handling.
-
-mod client_connect;
-mod client_disconnect;
-mod heartbeat;
-mod heartbeat_ack;
-mod hello;
-mod identify;
-mod ready;
-mod resume;
-mod select_protocol;
-mod session_description;
-mod speaking;
-
-pub use self::client_connect::ClientConnect;
-pub use self::client_disconnect::ClientDisconnect;
-pub use self::heartbeat::Heartbeat;
-pub use self::heartbeat_ack::HeartbeatAck;
-pub use self::hello::Hello;
-pub use self::identify::Identify;
-pub use self::ready::Ready;
-pub use self::resume::Resume;
-pub use self::select_protocol::SelectProtocol;
-pub use self::session_description::SessionDescription;
-pub use self::speaking::Speaking;
